+++ adapter/persistence/src/test/java/com/di/hostme/service/repository/city/JooqCityRepositoryTestSupport.java
@@ -2,6 +2,7 @@ package com.di.hostme.service.repository.city;
 
 import com.di.hostme.service.domain.city.entity.CityEntity;
 import com.di.hostme.service.domain.city.query.ImmutableFindCityByIdQuery;
+import com.di.hostme.service.domain.city.query.ImmutableListCityQuery;
 import java.util.List;
 import java.util.UUID;
 
@@ -11,7 +12,7 @@ public interface JooqCityRepositoryTestSupport {
     return repository.findOne(ImmutableFindCityByIdQuery.of(id, countryId));
   }
 
-  default List<CityEntity> listCities(JooqCityRepository repository) {
-    return repository.list();
+  default List<CityEntity> listCities(JooqCityRepository repository, UUID countryId) {
+    return repository.list(ImmutableListCityQuery.of(countryId));
   }
 }