+++ app/src/test/java/com/di/hostme/service/city/CityListFeatureSpec.java
@@ -3,6 +3,7 @@ package com.di.hostme.service.city;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.di.hostme.service.rest.api.dto.city.CityListResponse;
+import java.util.UUID;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
@@ -24,5 +25,15 @@ public class CityListFeatureSpec extends CityFeatureSpec {
 
     assertThat(listResponse.getCities()).isNotEmpty();
     assertThat(listResponse.getCities().size()).isEqualTo(3);
+  }
+
+  @Test
+  public void thatEmptyCityListForNonExistingCountryIdCanBeRetrieved() {
+    final UUID notExistingCountryId = UUID.randomUUID();
+
+    final CityListResponse listResponse =
+        listCities(notExistingCountryId, HttpStatus.OK, CityListResponse.class);
+
+    assertThat(listResponse.getCities()).isEmpty();
   }
 }